Task 3: Raffle logic and improvements.
A new endpoint was created for having the raffle. Token authentication was added to it. A new model
with a relationship of one to one with person model was created and migrated to represent each raffle.
Improvements was made on base controller and person tests. All other classes related to the base controller
and its respective tests was updated accordingly. Many of these improvements was made based on suggested
changes from review and pairing with Halan. Some of those changes was suggested on the review of task-2,
but it were only in the commit of task-3.